fix(remix-node): don't delete entire session directory when destroying empty session file (#3113)

+++ packages/remix-node/sessions/fileStorage.ts
@@ -88,6 +88,11 @@ export function createFileSessionStorage({
       await fsp.writeFile(file, content, "utf-8");
     },
     async deleteData(id) {
+      // Return early if the id is empty, otherwise we'll end up trying to
+      // unlink the dir, which will cause the EPERM error.
+      if (!id) {
+        return;
+      }
       try {
         await fsp.unlink(getFile(dir, id));
       } catch (error: any) {
